+++ firmware/prototyper/src/sbi/mod.rs
@@ -114,11 +114,17 @@ impl runtime::rustsbi::RustSBI for PlatformPolicy {
     ) -> runtime::rustsbi::SbiRet {
         use crate::riscv::allwinner_v821 as andes;
         use runtime::rustsbi::SbiRet;
+        if extension == andes::AWBASE_EXTENSION {
+            return andes::handle_awbase(function);
+        }
         if extension == andes::EXTENSION {
             return andes::handle(function, param, crate::platform::supervisor_memory());
         }
         if extension == 0x10 && function == 3 && param[0] == andes::EXTENSION {
             return SbiRet::success(usize::from(andes::available()));
+        }
+        if extension == 0x10 && function == 3 && param[0] == andes::AWBASE_EXTENSION {
+            return SbiRet::success(usize::from(andes::awbase_available()));
         }
         SBI_DISPATCHER
             .get()